Fix fence behavior in H.264 encoder.
This was all broken, but happened to work when all the work was done
on a (single-stream) GPU anyway. It broke horribly when the CPU actually
needed to look at the data (for non-zerocopy H.264 encoding).

This is, as far as I know, the last fix needed to run Nageru on NVIDIA
GPUs (tested with a GTX 950). Encoding is still done by Quick Sync,
not NVENC.