Signal the timestamps of the next packet to be written before flushing
Muxers that support this (e.g. movenc) can use it to write better
durations for the last packet in a fragment.

This requires checking the flag, since zero sized packets could be
a valid thing to write to the stream for other muxers, and aren't
filtered out at the libavformat generic code level.